refactor rest logic
